+++ packages/camera/camera_android_camerax/lib/src/android_camera_camerax_flutter_api_impls.dart
@@ -6,6 +6,7 @@ import 'camera_info.dart';
 import 'camera_selector.dart';
 import 'camerax_library.pigeon.dart';
 import 'java_object.dart';
+import 'process_camera_provider.dart';
 
 /// Handles initialization of Flutter APIs for the Android CameraX library.
 class AndroidCameraXCameraFlutterApis {
@@ -14,6 +15,7 @@ class AndroidCameraXCameraFlutterApis {
     JavaObjectFlutterApiImpl? javaObjectFlutterApi,
     CameraInfoFlutterApiImpl? cameraInfoFlutterApi,
     CameraSelectorFlutterApiImpl? cameraSelectorFlutterApi,
+    ProcessCameraProviderFlutterApiImpl? processCameraProviderFlutterApi,
   }) {
     this.javaObjectFlutterApi =
         javaObjectFlutterApi ?? JavaObjectFlutterApiImpl();
@@ -21,6 +23,8 @@ class AndroidCameraXCameraFlutterApis {
         cameraInfoFlutterApi ?? CameraInfoFlutterApiImpl();
     this.cameraSelectorFlutterApi =
         cameraSelectorFlutterApi ?? CameraSelectorFlutterApiImpl();
+    this.processCameraProviderFlutterApi = processCameraProviderFlutterApi ??
+        ProcessCameraProviderFlutterApiImpl();
   }
 
   static bool _haveBeenSetUp = false;
@@ -40,12 +44,17 @@ class AndroidCameraXCameraFlutterApis {
   /// Flutter Api for [CameraSelector].
   late final CameraSelectorFlutterApiImpl cameraSelectorFlutterApi;
 
+  /// Flutter Api for [ProcessCameraProvider].
+  late final ProcessCameraProviderFlutterApiImpl
+      processCameraProviderFlutterApi;
+
   /// Ensures all the Flutter APIs have been setup to receive calls from native code.
   void ensureSetUp() {
     if (!_haveBeenSetUp) {
       JavaObjectFlutterApi.setup(javaObjectFlutterApi);
       CameraInfoFlutterApi.setup(cameraInfoFlutterApi);
       CameraSelectorFlutterApi.setup(cameraSelectorFlutterApi);
+      ProcessCameraProviderFlutterApi.setup(processCameraProviderFlutterApi);
       _haveBeenSetUp = true;
     }
   }