+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.10.0
 
 * Allows camera to be switched while video recording.
+
+## 0.9.8+6
+
+* Updates code for `no_leading_underscores_for_local_identifiers` lint.
 * Updates minimum Flutter version to 2.10.
 
 ## 0.9.8+5