+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.4
+
+* Updates code for `no_leading_underscores_for_local_identifiers` lint.
+
 ## 2.2.3
 
 * Updates imports for `prefer_relative_imports`.